refactor(notes): isolate deterministic editor save coordination
Issue #21: extract FloorpNoteSaveCoordinator from the editor view
controller into its own file and move autosave scheduling into it with an
injectable sleep, so autosave, coalescing, lifecycle, retry, and
cancellation are testable without real timers or a presented controller.

- FloorpNoteSaveCoordinator.swift: change versions, in-flight save
  coalescing, retry state, and scheduleAutosave(delayNanoseconds:sleep:)
  with cancelAutosave() (nonisolated for deinit).
- FloorpNoteEditorViewController: delegates scheduling to the coordinator;
  onAutosave wires back to the save path. No behavior change.
- Four new deterministic autosave tests (fake-sleep gating, reschedule
  cancellation, cancel-before-fire, callback saves dirty draft).
- Targeted suite 58/58 green; full FloorpCI 477/477 green.
